[Feat] GraphQL API에 대한 Error Handling 추가
서버의 안정성을 위해 GraphQL API에 대한 Error Handling 추가.
