+++ api-server/api/graphql/mutations/PostMutation.js
@@ -2,9 +2,7 @@ const { GraphQLNonNull, GraphQLString, GraphQLID } = require('graphql');
 const { GraphQLUpload } = require('graphql-upload');
 
 const { PostType } = require('../types');
-const { Post } = require('../../../db');
-const { insertPost, updatePostInfo } = require('../../services/PostService');
-const s3 = require('../../../upload');
+const postService = require('../../services/post-service');
 
 const createPost = {
   type: PostType,
@@ -20,32 +18,10 @@ const createPost = {
       type: GraphQLString,
     },
   },
-  async resolve(parent, { file, userId, content }) {
-    try {
-      const postInfo = { userId, content };
-      const { filename, createReadStream } = await file;
-      const stream = createReadStream();
-
-      const extensionRegex = /(.jpg|.gif|.jpeg|.png)$/i;
-      if (!extensionRegex.test(filename)) {
-        return false;
-      }
-
-      const imageFile = await s3
-        .upload({
-          Bucket: `${process.env.BUCKET}`,
-          ACL: 'public-read',
-          Key: `post/${Date.now().toString()}_${filename}`,
-          Body: stream,
-        })
-        .promise();
-
-      const post = await insertPost(imageFile, postInfo);
-      return post;
-    } catch (e) {
-      console.log(e);
-      return e;
-    }
+  async resolve(_, { file, userId, content }) {
+    const postInfo = { userId, content };
+    const post = await postService.insertPost(file, postInfo);
+    return post;
   },
 };
 
@@ -59,14 +35,8 @@ const deletePost = {
     },
   },
   resolve: async (_, { postURL }) => {
-    await Post.destroy({
-      where: {
-        postURL,
-      },
-    });
-    return {
-      postURL,
-    };
+    await postService.removePost(postURL);
+    return { postURL };
   },
 };
 
@@ -83,7 +53,10 @@ const updatePost = {
       type: new GraphQLNonNull(GraphQLID),
     },
   },
-  resolve: async (_, { id, content }) => await updatePostInfo(content, id),
+  resolve: async (_, { id, content }) => {
+    await postService.updatePostInfo(content, id);
+    return { id, content };
+  },
 };
 
 module.exports = {
