+++ api-server/api/graphql/queries/CommentQuery.js
@@ -1,7 +1,7 @@
 const { GraphQLInt, GraphQLID, GraphQLList } = require('graphql');
 
 const { CommentType } = require('../types');
-const { Comment } = require('../../../db');
+const { getCommentList } = require('../../services/comment-service');
 
 const commentQuery = {
   type: new GraphQLList(CommentType),
@@ -10,13 +10,9 @@ const commentQuery = {
     limit: { type: GraphQLInt },
     offset: { type: GraphQLInt },
   },
-  resolve: (post, args) => {
-    return Comment.findAll({
-      where: { PostId: args.PostId },
-      offset: args.offset,
-      limit: args.limit,
-      order: [['updatedAt', 'DESC']],
-    });
+  resolve: async (_, { PostId, offset, limit }) => {
+    const commentList = await getCommentList(PostId, offset, limit);
+    return commentList;
   },
 };
