+++ api-server/api/graphql/queries/FollowListQuery.js
@@ -6,7 +6,7 @@ const {
   getFollowIdList,
   getFollowDataList,
   getFollowUpdatedDataList,
-} = require('../../services/FollowListService');
+} = require('../../services/followlist-service');
 
 const followListQuery = {
   type: new GraphQLList(UserType),
@@ -19,19 +19,14 @@ const followListQuery = {
     },
   },
   resolve: async (_, args) => {
-    try {
-      const followList = await getFollowList(args);
-      const followIdList = getFollowIdList(followList);
-      const followDataList = await getFollowDataList(followIdList);
-      const followUpdatedDataList = await getFollowUpdatedDataList(
-        followDataList,
-        args,
-      );
-      return followUpdatedDataList;
-    } catch (err) {
-      console.log(err);
-      return { error: err.message };
-    }
+    const followList = await getFollowList(args);
+    const followIdList = getFollowIdList(followList);
+    const followDataList = await getFollowDataList(followIdList);
+    const followUpdatedDataList = await getFollowUpdatedDataList(
+      followDataList,
+      args,
+    );
+    return followUpdatedDataList;
   },
 };
