+++ api-server/api/graphql/queries/LikerListQuery.js
@@ -1,7 +1,7 @@
 const { GraphQLList, GraphQLID } = require('graphql');
 
 const { LikerType } = require('../types');
-const { getLikerList } = require('../../services/PostService');
+const { getLikerList } = require('../../services/post-like-service');
 
 const likerListQuery = {
   type: new GraphQLList(LikerType),
@@ -11,12 +11,8 @@ const likerListQuery = {
     },
   },
   resolve: async (_, { postId }) => {
-    try {
-      const likerList = await getLikerList(postId);
-      return likerList;
-    } catch (err) {
-      return { error: err.message };
-    }
+    const likerList = await getLikerList(postId);
+    return likerList;
   },
 };
