+++ api-server/api/graphql/queries/SearchHashtagQuery.js
@@ -1,20 +1,16 @@
 const { GraphQLString, GraphQLList } = require('graphql');
 
 const { SearchHashtagType } = require('../types');
-const { HashTag, Sequelize } = require('../../../db');
-
-const { Op } = Sequelize;
+const { getHashTags } = require('../../services/hashtag-service');
 
 const searchHashtagQuery = {
   type: new GraphQLList(SearchHashtagType),
   args: {
     id: { type: GraphQLString },
   },
-  resolve: (hashtag, args) => {
-    return HashTag.findAll({
-      where: { name: { [Op.like]: `%${args.id}%` } },
-      limit: 10,
-    });
+  resolve: async (_, { id }) => {
+    const hashTags = await getHashTags(id);
+    return hashTags;
   },
 };
