+++ api-server/api/graphql/queries/SearchUserQuery.js
@@ -1,25 +1,16 @@
 const { GraphQLString, GraphQLList } = require('graphql');
 
 const { SearchUserType } = require('../types');
-const { User, Sequelize } = require('../../../db');
-
-const { Op } = Sequelize;
+const { searchUsers } = require('../../services/search-service');
 
 const searchUserQuery = {
   type: new GraphQLList(SearchUserType),
   args: {
     id: { type: GraphQLString },
   },
-  resolve: (user, args) => {
-    return User.findAll({
-      where: {
-        [Op.or]: [
-          { username: { [Op.like]: `%${args.id}%` } },
-          { name: { [Op.like]: `%${args.id}%` } },
-        ],
-      },
-      limit: 10,
-    });
+  resolve: async (_, { id }) => {
+    const users = await searchUsers(id);
+    return users;
   },
 };
