+++ api-server/api/graphql/queries/TaggedPostsQuery.js
@@ -1,11 +1,7 @@
 const { GraphQLString, GraphQLInt, GraphQLList } = require('graphql');
 
 const { PostType } = require('../types');
-const {
-  Sequelize: { Op },
-  Post,
-  UserTag,
-} = require('../../../db');
+const { getTaggedPosts } = require('../../services/post-service');
 
 const taggedPostsQuery = {
   type: new GraphQLList(PostType),
@@ -13,25 +9,9 @@ const taggedPostsQuery = {
     username: { type: GraphQLString },
     myId: { type: GraphQLInt },
   },
-  resolve: async (_, { username, myId }) => {
-    try {
-      const taggedPosts = await UserTag.findAll({
-        attributes: ['postId'],
-        where: { username },
-      });
-      const taggedPostIdList = taggedPosts.map(
-        taggedPost => taggedPost.dataValues.postId,
-      );
-      const taggedPostsInfo = await Post.findAll({
-        attributes: ['postURL', 'imageURL'],
-        where: { id: { [Op.in]: taggedPostIdList } },
-        order: [['updatedAt', 'DESC']],
-      });
-      return taggedPostsInfo;
-    } catch (e) {
-      console.log(e.message);
-      return { error: e.message };
-    }
+  resolve: async (_, { username }) => {
+    const taggedPosts = await getTaggedPosts(username);
+    return taggedPosts;
   },
 };
