+++ api-server/api/graphql/queries/UserPageQuery.js
@@ -1,7 +1,7 @@
 const { GraphQLString, GraphQLID } = require('graphql');
 
 const { UserPageType } = require('../types');
-const { getUserPageData } = require('../../services/UserPageService');
+const { getUserPageData } = require('../../services/userpage-service');
 
 const userPageQuery = {
   type: UserPageType,
@@ -10,13 +10,8 @@ const userPageQuery = {
     myId: { type: GraphQLID },
   },
   resolve: async (_, args) => {
-    try {
-      const data = await getUserPageData(args);
-      return data;
-    } catch (e) {
-      console.log(e.message);
-      return { error: e.message };
-    }
+    const data = await getUserPageData(args);
+    return data;
   },
 };
