+++ api-server/api/services/userpage-service.js
@@ -1,4 +1,5 @@
 const { Post, User, UserFollow } = require('../../db');
+const { errorName } = require('../../error');
 
 const getUserInfo = async username => {
   const userInfo = await User.findOne({
@@ -43,26 +44,30 @@ const getPostCard = async userId => {
 };
 
 const getUserPageData = async args => {
-  const { username, myId } = args;
-  let userInfo = await getUserInfo(username);
-  const isExistingUser = !!userInfo;
-  let postCard = [];
-  if (isExistingUser) {
-    const userId = userInfo.id;
-    const isFollowing = await checkFollowing(userId, myId);
-    const followersNum = await getFollowersNum(userId);
-    const followsNum = await getFollowsNum(userId);
-    postCard = await getPostCard(userId);
-    userInfo = {
-      ...userInfo.dataValues,
-      isFollowing,
-      postNumber: postCard.length,
-      followersNum,
-      followsNum,
-    };
+  try {
+    const { username, myId } = args;
+    let userInfo = await getUserInfo(username);
+    const isExistingUser = !!userInfo;
+    let postCard = [];
+    if (isExistingUser) {
+      const userId = userInfo.id;
+      const isFollowing = await checkFollowing(userId, myId);
+      const followersNum = await getFollowersNum(userId);
+      const followsNum = await getFollowsNum(userId);
+      postCard = await getPostCard(userId);
+      userInfo = {
+        ...userInfo.dataValues,
+        isFollowing,
+        postNumber: postCard.length,
+        followersNum,
+        followsNum,
+      };
+    }
+    const data = { isExistingUser, userInfo, postCard };
+    return data;
+  } catch (err) {
+    throw new Error(errorName.USER_PAGE_QUERY_ERROR);
   }
-  const data = { isExistingUser, userInfo, postCard };
-  return data;
 };
 
 module.exports = { getUserPageData };
