+++ api-server/package.json
@@ -33,6 +33,7 @@
     "cors": "^2.8.5",
     "cross-env": "^6.0.3",
     "dotenv": "^8.2.0",
+    "easygraphql-format-error": "0.0.3",
     "express": "~4.16.1",
     "express-graphql": "^0.9.0",
     "express-session": "^1.17.0",
